# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "haupham404"
        },
        "version": "1.0"
    },
    "host": "localhost:1909",
    "paths": {
        "/addmovies": {
            "post": {
                "description": "login username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonggoDB"
                ],
                "summary": "Putmongo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/adduser": {
            "post": {
                "description": "Add new user to database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonggoDB"
                ],
                "summary": "AddUser",
                "parameters": [
                    {
                        "type": "string",
                        "default": "application/json",
                        "description": "application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1000",
                        "description": "1000",
                        "name": "Content-Length",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "localhost",
                        "description": "localhost",
                        "name": "Host",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "model.RegisterUser",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/firebase/getfirebase": {
            "get": {
                "description": "getfirebase",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Firebase"
                ],
                "summary": "Getfirebase",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/firebase/putfirebase": {
            "post": {
                "description": "putfirebase",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Firebase"
                ],
                "summary": "Putfirebase",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/getmovies": {
            "get": {
                "description": "login username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonggoDB"
                ],
                "summary": "GetMovies",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Movies"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/gormdb/adduser": {
            "post": {
                "description": "Add new user to database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gormDB"
                ],
                "summary": "AddUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorizationc",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "model.RegisterUser",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gormdb/login": {
            "post": {
                "description": "login username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gormDB"
                ],
                "summary": "LoginUser",
                "parameters": [
                    {
                        "description": "model.Login",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Login"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Token"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/gormdb/user": {
            "get": {
                "description": "get username from token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gormDB"
                ],
                "summary": "GetUser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "login username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonggoDB"
                ],
                "summary": "LoginUser",
                "parameters": [
                    {
                        "type": "string",
                        "default": "application/json",
                        "description": "application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1000",
                        "description": "1000",
                        "name": "Content-Length",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "localhost",
                        "description": "localhost",
                        "name": "Host",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "model.Login",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Login"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Token"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "description": "GetLogout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonggoDB"
                ],
                "summary": "GetLogout",
                "parameters": [
                    {
                        "type": "string",
                        "default": "application/json",
                        "description": "application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1000",
                        "description": "1000",
                        "name": "Content-Length",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "localhost",
                        "description": "localhost",
                        "name": "Host",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/message": {
            "post": {
                "description": "GetMessage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonggoDB"
                ],
                "summary": "GetMessage",
                "parameters": [
                    {
                        "type": "string",
                        "default": "application/json",
                        "description": "application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1000",
                        "description": "1000",
                        "name": "Content-Length",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "localhost",
                        "description": "localhost",
                        "name": "Host",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "model.InputGetMessage",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputGetMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        },
        "/savemessage": {
            "post": {
                "description": "SaveMessage",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MonggoDB"
                ],
                "summary": "SaveMessage",
                "parameters": [
                    {
                        "type": "string",
                        "default": "application/json",
                        "description": "application/json",
                        "name": "Content-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1000",
                        "description": "1000",
                        "name": "Content-Length",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "localhost",
                        "description": "localhost",
                        "name": "Host",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "model.Message",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "model.InputGetMessage": {
            "type": "object",
            "properties": {
                "receiver": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                }
            }
        },
        "model.Login": {
            "type": "object",
            "properties": {
                "Password": {
                    "type": "string"
                },
                "UserID": {
                    "type": "string"
                }
            }
        },
        "model.Message": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "receiver": {
                    "type": "string"
                },
                "sender": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
        "model.Movies": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.RegisterUser": {
            "type": "object",
            "properties": {
                "Password": {
                    "type": "string"
                },
                "UserID": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "model.Token": {
            "type": "object",
            "properties": {
                "Authorization": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    }
}